// src/lib/types.ts
export interface Token {
  id: string;
  name: string;
  ticker: string;
  ca: string;
  devWallet: string | null;
  mcap: number;
  mcapBnb: number;
  price: number;
  createdAt: number;
  holders: number;
  image: string;
  change24h: number;
  x: number;
  y: number;
  vx: number;
  vy: number;
  activityBoost: number;
  devHoldPercent: number;
  burnPercent: number;
  sniperHoldPercent: number;
  website: string | null;
  twitter: string | null;
  telegram: string | null;
  bondingCurve: boolean;
  bondProgress: number;
  reserveBnb: number;
  graduated: boolean;
  listed: boolean;
  taxRate: number;
  taxEarned: number;
  beneficiary: string | null;
  dexPaid: boolean;
  dexPairCount: number;
  aveLogo: boolean;
  description: string | null;
  section: string;
  lastMcap?: number;
  newlyDetectedAt?: number;
  volume24h: number;
  liquidity: number;
  buys24h: number;
  sells24h: number;
  dexUrl: string | null;
  isPartner?: boolean;
}

export interface RawToken {
  address: string;
  name?: string;
  ticker?: string;
  mcap?: number;
  mcapBnb?: number;
  price?: number;
  createdAt?: string;
  holders?: number;
  image?: string;
  change24h?: number;
  devHoldPercent?: number;
  burnPercent?: number;
  devWallet?: string;
  sniperHoldPercent?: number;
  website?: string;
  twitter?: string;
  telegram?: string;
  bondingCurve?: boolean;
  bondProgress?: number;
  reserveBnb?: number;
  graduated?: boolean;
  listed?: boolean;
  taxRate?: number;
  taxEarned?: number;
  beneficiary?: string;
  dexPaid?: boolean;
  dexPairCount?: number;
  aveLogo?: boolean;
  description?: string;
  section?: string;
  newlyDetectedAt?: number;
  volume24h?: number;
  liquidity?: number;
  buys24h?: number;
  sells24h?: number;
  dexUrl?: string;
  isPartner?: boolean;
}